Made some changes to the Code
I've made some to the Holberton code. The changes are as follows, line 6 "Holberton" to "Hello Holberton", and line 4 to modify the 10 into a 32. I honestly would of said "howdy" or "heya" instead of hello haha, but I'm thinking you guys are trying to figure out if I can follow directions first before I can become an individual. But yeah, this seemed pretty fun wish it was a little trickier I love trying to figure how to do things out, this was kinda written in instructions.